EES-823 Extract NotFoundPage and ErrorPage from _error.tsx
