+++ src/explore-education-statistics-frontend/src/pages/404.tsx
@@ -1,0 +1,3 @@
+import NotFoundPage from '@frontend/modules/NotFoundPage';
+
+export default NotFoundPage;
